Move Square class into its own file
